% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{confint.gce}
\alias{confint.gce}
\title{Confidence Intervals for gce Model Parameters}
\usage{
\method{confint}{gce}(object, parm, level = 0.95, method = "basic")
}
\arguments{
\item{object}{Fitted \code{gce} model object.}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing,
all parameters are considered.}

\item{level}{the confidence level required.}

\item{method}{.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each parameter. These will be labelled as (1-level)/2 and
1 - (1-level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Computes bootstrap confidence intervals for one or more parameters in a gce
fitted model.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    ci = 95,
    verbose = FALSE)

confint(res_gce_package)
confint(res_gce_package, parm = c("X001"), level = 0.99)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
