% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{deviance.gce}
\alias{deviance.gce}
\title{Model Deviance}
\usage{
\method{deviance}{gce}(object)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\value{
The value of the deviance extracted from a \code{gce} object.
}
\description{
Returns the deviance of a fitted model \code{gce} object.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

deviance(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
