% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{df.residual.gce}
\alias{df.residual.gce}
\title{Residual Degrees-of-Freedom}
\usage{
\method{df.residual}{gce}(object)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\value{
The value of the residual degrees-of-freedom extracted from a gce
object.
}
\description{
Returns the residual degrees-of-freedom extracted from a fitted model gce
object.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

df.residual(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
