% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{fitted.gce}
\alias{fitted.gce}
\title{Calculate gce Fitted Values}
\usage{
\method{fitted}{gce}(object)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\value{
Returns a vector with the fitted values for the linear model
represented by a gce object.
}
\description{
The fitted values for the linear model represented by a gce object are
extracted.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

fitted(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
