% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{formula.gce}
\alias{formula.gce}
\title{Extract Model Formula from gce object}
\usage{
\method{formula}{gce}(object)
}
\arguments{
\item{object}{fitted \code{gce} object.}
}
\description{
Returns the model used to fit gce object.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

formula(res_gce_package)
# y ~ .

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
