% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_k3_yk2_cn1_df.R
\docType{data}
\name{gce_k3_yk2_cn1_df}
\alias{gce_k3_yk2_cn1_df}
\title{Simulated data set generated with gdata_GCE}
\format{
A \code{data.frame} containing:
\describe{
  \item{X001}{A N(0,1) independent variable.}
  \item{X002}{A N(0,1) independent variable.}
  \item{X003}{A N(0,1) independent variable.}
  \item{X004}{A N(0,1) independent variable.}
  \item{X005}{A N(0,1) independent variable.}
  \item{y}{A Dependent variable}
}
}
\usage{
gce_k3_yk2_cn1_df
}
\description{
Simulated data, used to demonstrate the features of gce
}
\examples{
# gce_k3_yk2_cn1_df <-
#   gdata_GCE(
#     n = 100,
#     cont.k = 3,
#     y.gen.cont.k = 2,
#     y.gen.cont.beta = c(5, 10),
#     intercept.beta = 0,
#     Xgenerator = "svd",
#     condnumber = 1,
#     mu = 0,
#     sd = 1,
#     error_dist = "normal",
#     error_normal_mean = 0,
#     error.dist.snr = 5,
#     dataframe = TRUE,
#     seed = 230676
#   )
data(gce_k3_yk2_cn1_df)
}
\keyword{datasets}
