% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_standardized_coef.R
\name{max_standardized_coef}
\alias{max_standardized_coef}
\title{Estimated maximum absolute regression standardized coefficient}
\usage{
max_standardized_coef(k, maxcor, percentile)
}
\arguments{
\item{k}{Number of parameters}

\item{maxcor}{Maximum absolute correlation}

\item{percentile}{Percentile}
}
\value{
Returns an \code{percentile} based estimated maximum absolute value
for a standardized regression coefficient given a linear regression model
with \code{k} parameters and a maximum correlation value of \code{maxcor}.
}
\description{
Returns the estimated maximum absolute regression standardized coefficient
}
\examples{
max_standardized_coef(k = 3, maxcor = .70, percentile = 90)
# [1] 1.024147

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
