% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{plot.gce}
\alias{plot.gce}
\title{Plot Diagnostics for a gce Object}
\usage{
\method{plot}{gce}(
  object,
  which = c(1, 2, 3, 4),
  prior = TRUE,
  caption = list("Residuals vs Fitted", "Q-Q Residuals", "Scale-Location",
    "Error vs supports"),
  panel = {
     if (add.smooth) {
         function(x, y, ...) {
            
    panel.smooth(x, y, iter = iter.smooth, ...)
         }
     }
     else points
 },
  sub.caption = NULL,
  main = "",
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  id.n = 3,
  labels.id = names(residuals(object)),
  cex.id = 0.75,
  qqline = TRUE,
  add.smooth = getOption("add.smooth"),
  iter.smooth = 3,
  label.pos = c(4, 2),
  cex.caption = 1,
  cex.oma.main = 1.25,
  extend.ylim.f = 0.08,
  ...
)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\description{
Three plots are currently available: a plot of observed against fitted
values, a Scale-Location plot of residuals against fitted values and a Q-Q
plot of residuals.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    ci = 95,
    verbose = FALSE)

plot(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
