% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{print.gce}
\alias{print.gce}
\title{Print gce}
\usage{
\method{print}{gce}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{fitted \code{gce} object.}

\item{digits}{significant digits in printout.}

\item{...}{additional print arguments.}
}
\value{
A small summary of a \code{gce} object is returned.
}
\description{
Print gce
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

res_gce_package
print(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
