% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{print.summary.gce}
\alias{print.summary.gce}
\title{Print Summary of gce Model Fits}
\usage{
\method{print}{summary.gce}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}

\item{digits}{The number of significant digits to use when printing..}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function returns a list of summary statistics of the
 fitted linear model given in object, using the components (list elements)
 "call" and "terms" from its argument, plus...
}
\description{
\code{print.summary} method for class "gce".
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    ci = 95,
    verbose = FALSE)

summary(res_gce_package)
summary(res_gce_package, ci.level = 0.99, ci.method = "basic")

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
