% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{residuals.gce}
\alias{residuals.gce}
\title{Extract gce Model Residuals}
\usage{
\method{residuals}{gce}(object, type = c("working", "response", "deviance", "pearson"))
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\value{
Returns the residuals from a gce object
}
\description{
\code{residuals} is a function which extracts model residuals from gce
objects.
The abbreviated form \code{resid} is an alias for \code{residuals}.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

residuals(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
