% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_back_coef.R
\name{scale_back_coef}
\alias{scale_back_coef}
\title{Scale coefficients back}
\usage{
scale_back_coef(X.scaled, y.scaled, betas.scaled, intercept = TRUE)
}
\arguments{
\item{X.scaled}{A matrix.}

\item{y.scaled}{A vector.}

\item{betas.scaled}{A vector of given scaled coefficients.}

\item{intercept}{logical indicating if}
}
\value{
Returns a vector of unscaled numeric regression coefficients.
}
\description{
Returns the unscaled regression coefficients
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
