% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{summary.gce}
\alias{summary.gce}
\title{Summarise a linear regression model via generalized cross entropy fit}
\usage{
\method{summary}{gce}(object, ci.level = 0.95, ci.method = "basic")
}
\arguments{
\item{object}{Fitted \code{gce} model object.}

\item{ci.level}{the confidence level required to compute a bootstrap
confidence interval.}

\item{ci.method}{the method used to compute a bootstrap a confidence interval.}
}
\value{
The function computes and returns a list of summary statistics of the
 fitted linear model given in \code{object}, using the components
 (list elements) "call" and "terms" from its argument, plus...
}
\description{
Generic function used to produce summary information from a fitted linear
regression model via generalized cross entropy as represented by
\code{object} of class \code{gce}.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    ci = 95,
    verbose = FALSE)

sm_res_gce_package <- summary(res_gce_package)
str(sm_res_gce_package)
sm_res_gce_package_99 <- summary(res_gce_package, ci.level = 0.99,
 ci.method = "basic")
sm_res_gce_package_99$coefficients
}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
