% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce_family.R
\name{variable.names.gce}
\alias{variable.names.gce}
\title{Variable Names of gce Fitted Models}
\usage{
\method{variable.names}{gce}(object)
}
\arguments{
\item{object}{Fitted \code{gce} model object.}
}
\value{
A character vector.
}
\description{
Simple utility returning variable names.
}
\examples{
res_gce_package <-
  gce(y~.,
    data = gce_k3_yk2_cn1_df,
    model = TRUE,
    method = "dual",
    verbose = FALSE)

variable.names(res_gce_package)

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
