% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdata_GCE.R
\name{gdata_GCE}
\alias{gdata_GCE}
\title{Data generating function}
\usage{
gdata_GCE(
  n,
  bin.k = 0,
  bin.prob = c(0.5, 0.3, 0.8),
  cont.k = 5,
  y.gen.bin.k = 0,
  y.gen.bin.beta = c(1, 2),
  y.gen.bin.prob = c(0.1, 0.2),
  y.gen.cont.k = 5,
  y.gen.cont.beta = c(2, 4, 6, 8, 10),
  y.gen.cont.mod.k = 0,
  y.gen.cont.mod.beta = matrix(c(-2, 2), 1, 2, byrow = TRUE),
  y.gen.bin.mod.prob = c(0.5),
  y.gen.cont.sp.k = 0,
  y.gen.cont.sp.groups = 2,
  y.gen.cont.sp.rho = 0.2,
  y.gen.cont.sp.dif = 1,
  intercept.beta = 0,
  Xgenerator.method = "simstudy",
  Xgenerator.seed = NULL,
  corMatrix = 100,
  corstr = "ind",
  condnumber = 1,
  mu = 0,
  muvect = NULL,
  sd = 1,
  sdvect = NULL,
  rho = 0,
  error.dist = "normal",
  error.dist.mean = 0,
  error.dist.sd = 1,
  error.dist.snr = NULL,
  error.dist.df = 2,
  dataframe = FALSE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of individuals.}

\item{bin.k}{Number of binary variables.}

\item{bin.prob}{A vector of probabilities with length equal to \code{bin.k}.}

\item{cont.k}{Number of continuous variables.}

\item{y.gen.bin.k}{.}

\item{y.gen.bin.beta}{.}

\item{y.gen.bin.prob}{.}

\item{y.gen.cont.k}{Number continuous variables used for generating y. If NULL
then \code{usercoeff} must be supplied.}

\item{y.gen.cont.beta}{.}

\item{y.gen.cont.mod.k}{.}

\item{y.gen.cont.mod.beta}{.}

\item{y.gen.bin.mod.prob}{.}

\item{y.gen.cont.sp.k}{.}

\item{y.gen.cont.sp.groups}{.}

\item{y.gen.cont.sp.rho}{.}

\item{y.gen.cont.sp.dif}{.}

\item{intercept.beta}{.}

\item{Xgenerator.method}{Method used to generate X data ( \code{"simstudy"}
or \code{"svd"}).}

\item{Xgenerator.seed}{A seed used when \code{Xgenerator} is \code{"svd"}.}

\item{corMatrix}{A value for d, NULL or a correlation matrix to be used when
\code{Xgenerator} is \code{"simstudy"}.}

\item{corstr}{correlation structure (\code{"ind"}, \code{"cs"} or
\code{"ar1"}) (see \code{\link[simstudy]{genCorData}}) to be used when
\code{Xgenerator} is \code{"simstudy"} and \code{corMatrix} is NULL.}

\item{condnumber}{A value for the condition number of the X matrix to be used
when \code{Xgenerator} is \code{"svd"}.}

\item{mu}{The mean of the variables. To be used when all variables have the
same mean.}

\item{muvect}{A vector of means. The length of \code{muvect} must be \code{k}.}

\item{sd}{Standard deviation of the variables. To be used when all variables
have the same standard deviation.}

\item{sdvect}{A vector of standard deviations. The length of \code{sdvect}
must be \code{k}.}

\item{rho}{Correlation coefficient, \code{-1 <= rho <= 1}. Use if
\code{corMatrix} is NULL.}

\item{error.dist}{Distribution of the error. \code{"normal"} for nomal
distribution or \code{"t"} for t-student distribution.}

\item{error.dist.mean}{Mean value used when \code{error_dist} is
\code{"normal"}.}

\item{error.dist.sd}{Standard deviation value used when \code{error_dist} is
\code{"normal"}.}

\item{error.dist.snr}{.}

\item{error.dist.df}{Degrees of freedom used when \code{error_dist} is
\code{"t"}.}

\item{dataframe}{Logical. If \code{TRUE} returns a \code{data.frame} else
returns a \code{list}.}

\item{seed}{A seed for reproducibility.}
}
\value{
List composed of a vector of the dependent variable values (y), a
matrix of independent variables values (X), a vector of coefficient values
(coefficients) and a vector of the error values (epsilon).
}
\description{
Generates data
}
\examples{
# The following code generates a matrix X composed of 100 individuals (lines)
# and 10 independent variables (columns) generated from a normal distribution
# with mean 1 and standard deviation 0.
# The Xgenerator is "svd" using a condition number equal to 5.
# The y vector is generated using 2 of the generated independent variables
# multiplied by the coefficients 5 and 10.
# An error generated from a normal distribution with mean 0 and standard
# deviation obtained from a signal-to-noise ratio of 5 is added to the
# previous y vector.

gce_k3_yk2_cn1_list <- gdata_GCE(
  n = 100, cont.k = 3,
  y.gen.cont.k = 2, y.gen.cont.beta = c(5,10),
  intercept.beta = 0,
  Xgenerator.method = "svd", condnumber = 1,
  mu = 0, sd = 1,
  error.dist = "normal", error.dist.mean = 0, error.dist.snr = 5,
  seed = 230676)

gce_k3_yk2_cn1_list$coefficients
# [1]  0  0  0  0  5 10

gce_k3_yk2_cn1_df <- gdata_GCE(
  n = 100, cont.k = 3,
  y.gen.cont.k = 2, y.gen.cont.beta = c(5,10),
  intercept.beta = 0,
  Xgenerator.method = "svd", condnumber = 1,
  mu = 0, sd = 1,
  error.dist = "normal", error.dist.mean = 0, error.dist.snr = 5,
  dataframe = TRUE, seed = 230676)

summary(gce_k3_yk2_cn1_df)
#           X001                X002                X003
# Min.   :-0.346510   Min.   :-0.233598   Min.   :-0.194100
# 1st Qu.:-0.064938   1st Qu.:-0.066407   1st Qu.:-0.058178
# Median :-0.007034   Median : 0.009406   Median : 0.007381
# Mean   :-0.010142   Mean   :-0.003627   Mean   : 0.006016
# 3rd Qu.: 0.060668   3rd Qu.: 0.057542   3rd Qu.: 0.065680
# Max.   : 0.197434   Max.   : 0.316639   Max.   : 0.264895
#           X004                X005                 y
# Min.   :-0.267914   Min.   :-0.233906   Min.   :-2.92736
# 1st Qu.:-0.068691   1st Qu.:-0.086896   1st Qu.:-0.81050
# Median : 0.007156   Median :-0.003676   Median :-0.13388
# Mean   : 0.002325   Mean   :-0.006584   Mean   :-0.05318
# 3rd Qu.: 0.061177   3rd Qu.: 0.057585   3rd Qu.: 0.62974
# Max.   : 0.237203   Max.   : 0.289297   Max.   : 2.84185

}
\author{
Jorge Cabral, \email{jorgecabral@ua.pt}
}
